# colqwen_fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ColQwen2.5 Embedding API",
    "description": "API for generating embeddings from images and queries",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Health check endpoint",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/info": {
      "get": {
        "summary": "Version",
        "description": "Version endpoint",
        "operationId": "version_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/patches": {
      "post": {
        "summary": "Get N Patches",
        "description": "Calculate number of patches for given image dimensions and spatial merge size\n\nArgs:\n    request: PatchRequest containing a list of dimensions with 'width' and 'height' keys",
        "operationId": "get_n_patches_patches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/embed/queries": {
      "post": {
        "summary": "Embed Queries",
        "description": "Generate embeddings for text queries",
        "operationId": "embed_queries_embed_queries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryEmbeddingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/embed/images": {
      "post": {
        "summary": "Embed Images",
        "description": "Generate embeddings for uploaded images + image-token boundaries.",
        "operationId": "embed_images_embed_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_embed_images_embed_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageEmbeddingBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_embed_images_embed_images_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_embed_images_embed_images_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageEmbeddingBatchResponse": {
        "properties": {
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/ImageEmbeddingItem"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "embeddings"
        ],
        "title": "ImageEmbeddingBatchResponse"
      },
      "ImageEmbeddingItem": {
        "properties": {
          "embedding": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Embedding"
          },
          "image_patch_start": {
            "type": "integer",
            "title": "Image Patch Start"
          },
          "image_patch_len": {
            "type": "integer",
            "title": "Image Patch Len"
          }
        },
        "type": "object",
        "required": [
          "embedding",
          "image_patch_start",
          "image_patch_len"
        ],
        "title": "ImageEmbeddingItem"
      },
      "PatchBatchResponse": {
        "properties": {
          "results": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "PatchBatchResponse"
      },
      "PatchRequest": {
        "properties": {
          "dimensions": {
            "items": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Dimensions"
          }
        },
        "type": "object",
        "required": [
          "dimensions"
        ],
        "title": "PatchRequest"
      },
      "QueryEmbeddingResponse": {
        "properties": {
          "embeddings": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "embeddings"
        ],
        "title": "QueryEmbeddingResponse"
      },
      "QueryRequest": {
        "properties": {
          "queries": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Queries"
          }
        },
        "type": "object",
        "required": [
          "queries"
        ],
        "title": "QueryRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}